# openapi/ghf-grant-knowledge-api.yaml
openapi: 3.1.0
info:
  title: GHF Grant Knowledge API
  version: 1.0.0
  description: >
    Dropbox-only v1 middleware for Custom GPT Actions. The GPT sends one grant
    question or a pasted grant application as query text. The server searches
    only the approved Grant Halliburton Foundation Dropbox folders, then returns
    compact source excerpts for the GPT to draft paste-ready grant answers.
servers:
  - url: https://your-github-username-ghf-grant-know.vercel.app
    description: Vercel production deployment.
security:
  - bearerAuth: []
paths:
  /health:
    get:
      operationId: health
      summary: Confirm the service is running.
      security: []
      responses:
        "200":
          description: Service status.
          content:
            application/json:
              schema:
                type: object
                required: [status, service, version]
                properties:
                  status:
                    type: string
                    const: ok
                  service:
                    type: string
                    const: GHF Grant Knowledge API
                  version:
                    type: string
                    const: 1.0.0
  /search:
    post:
      operationId: searchGrantSources
      x-openai-isConsequential: false
      summary: Find relevant Dropbox source excerpts for a grant question or pasted grant application.
      description: >
        Search approved GHF Dropbox folders for source excerpts relevant to a
        grant question or pasted application. The client sends query text and
        an API key. Folder scope, Dropbox credentials, and source selection are
        enforced by the server.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              required: [query]
              properties:
                query:
                  type: string
                  minLength: 1
                  maxLength: 100000
                  description: One grant question, multiple grant questions, or a pasted grant application/document.
                character_limit:
                  type: integer
                  minimum: 1
                  maximum: 50000
                  description: Optional final-answer character limit for the GPT to honor while drafting.
                max_results:
                  type: integer
                  minimum: 1
                  maximum: 10
                  default: 5
                  description: Optional number of source excerpts to return.
      responses:
        "200":
          description: Relevant source excerpts from approved Dropbox folders.
          content:
            application/json:
              schema:
                type: object
                required: [query, query_characters, source, searched_folders, results]
                properties:
                  query:
                    type: string
                    description: Compact echo of the submitted query.
                  query_characters:
                    type: integer
                  source:
                    type: string
                    const: dropbox
                  searched_folders:
                    type: array
                    items:
                      type: string
                    description: Server-enforced approved Dropbox folders.
                  results:
                    type: array
                    maxItems: 10
                    items:
                      $ref: "#/components/schemas/SearchResult"
        "400":
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /files/upload:
    post:
      operationId: uploadDropboxFile
      x-openai-isConsequential: true
      summary: Upload a confirmed file to an approved current Dropbox folder.
      description: >
        After explicit user confirmation, upload one supported file to the
        current grant library or current submitted-grants folder. The server
        selects the destination path and never overwrites an existing file.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [target, file_name, confirmed]
              properties:
                target:
                  type: string
                  enum: [current_grant_library, current_grants_submitted]
                file_name:
                  type: string
                  minLength: 1
                  maxLength: 180
                content_base64:
                  type: string
                  maxLength: 7000000
                  description: Base64 file bytes. Provide this or content_text, never both.
                content_text:
                  type: string
                  maxLength: 5000000
                  description: Plain text for a .txt, .md, or .csv file. Provide this or content_base64.
                confirmed:
                  type: boolean
                  const: true
                  description: True only after the user explicitly confirms this upload.
      responses:
        "201":
          description: File uploaded without overwriting an existing file.
          content:
            application/json:
              schema:
                type: object
                required: [status, target, file]
                properties:
                  status:
                    type: string
                    const: uploaded
                  target:
                    type: string
                    enum: [current_grant_library, current_grants_submitted]
                  file:
                    $ref: "#/components/schemas/DropboxFile"
        "400":
          description: Invalid file request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Writes are disabled or the path is restricted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          description: File exceeds the configured size limit.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /files/move-to-archive:
    post:
      operationId: moveDropboxFileToArchive
      x-openai-isConsequential: true
      summary: Move a confirmed current Dropbox file into its matching archive.
      description: >
        After explicit user confirmation, move one approved current-library or
        current-submitted file into its server-mapped archive. The request
        cannot supply or override the destination path.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [source_path, confirmed]
              properties:
                source_path:
                  type: string
                  minLength: 1
                  maxLength: 1000
                  description: Exact safe path of a file in an approved current folder.
                confirmed:
                  type: boolean
                  const: true
                  description: True only after the user explicitly confirms this move.
      responses:
        "200":
          description: File moved to its matching archive.
          content:
            application/json:
              schema:
                type: object
                required: [status, file]
                properties:
                  status:
                    type: string
                    const: moved_to_archive
                  file:
                    allOf:
                      - $ref: "#/components/schemas/DropboxFile"
                      - type: object
                        required: [from_path]
                        properties:
                          from_path:
                            type: string
        "400":
          description: Invalid move request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Writes are disabled or the path is restricted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API key
  responses:
    Unauthorized:
      description: Missing or invalid API key.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    DropboxFile:
      type: object
      required: [name, path]
      properties:
        name:
          type: string
        path:
          type: string
        id:
          type: [string, "null"]
        rev:
          type: [string, "null"]
        size:
          type: [integer, "null"]
    SearchResult:
      type: object
      required: [rank, title, source_file, path, source_path, source_folder, source_category, excerpt, confidence]
      properties:
        rank:
          type: integer
          minimum: 1
          description: Ranking order, with the best result first.
        title:
          type: string
          description: Source title without the file extension.
        source_file:
          type: string
        funder:
          type: [string, "null"]
        year:
          type: [string, "null"]
        path:
          type: string
          description: Safe Dropbox source path inside an approved folder.
        source_path:
          type: string
          description: Same safe Dropbox source path, included for GPT clarity.
        source_folder:
          type: string
          enum:
            - Test Current Grant Library
        source_category:
          type: string
          enum: [approved_grant_folder, grantwriting_resources]
        excerpt:
          type: string
          maxLength: 2000
        confidence:
          type: string
          enum: [high, medium, low]
        document_type:
          type: [string, "null"]
        character_count:
          type: integer
        notes:
          type: string
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
